feat: export hook useYAt
