+++ src/charts/line/index.ts
@@ -16,6 +16,7 @@ import { useLineChartDatetime } from './useDatetime';
 import { useLineChartPrice } from './usePrice';
 import { useLineChart } from './useLineChart';
 import { LineChartHoverTrap } from '../line/HoverTrap';
+import { useYAt } from './useYAt';
 
 export * from './Chart';
 export * from './ChartPath';
@@ -54,4 +55,5 @@ export const LineChart = Object.assign(_LineChart, {
   usePrice: useLineChartPrice,
   useChart: useLineChart,
   HoverTrap: LineChartHoverTrap,
+  useYAt: useYAt,
 });
