+++ src/charts/line/types.ts
@@ -29,3 +29,14 @@ export type YDomain = {
   min: number;
   max: number;
 };
+
+export type AtPoint =
+  | {
+      index: number;
+      value?: never;
+    }
+  | {
+      index?: never;
+      value: number;
+    }
+  | number;
